rename InterOpCollectionBuilder && update auxilia op
