+++ research/query_service/ir/compiler/src/main/java/com/alibaba/graphscope/common/IrPlan.java
@@ -255,7 +255,10 @@ public class IrPlan implements Closeable {
 
                 Optional<OpArg> aliasOpt = baseOp.getAlias();
                 if (aliasOpt.isPresent()) {
-                    irCoreLib.setAuxiliaAlias(ptrAuxilia, (FfiNameOrId.ByValue) aliasOpt.get().getArg(), true);
+                    FfiNameOrId.ByValue alias = (FfiNameOrId.ByValue) aliasOpt.get().getArg();
+                    // aliases generated by compiler all start with ~
+                    boolean isQueryGiven = !ArgUtils.isHiddenStr(alias.name);
+                    irCoreLib.setAuxiliaAlias(ptrAuxilia, (FfiNameOrId.ByValue) aliasOpt.get().getArg(), isQueryGiven);
                 }
                 return ptrAuxilia;
             }
