+++ research/query_service/ir/compiler/src/main/java/com/alibaba/graphscope/gremlin/plugin/processor/IrStandardOpProcessor.java
@@ -30,7 +30,7 @@ import com.alibaba.graphscope.common.client.*;
 import com.alibaba.graphscope.common.config.Configs;
 import com.alibaba.graphscope.common.config.PegasusConfig;
 import com.alibaba.graphscope.common.intermediate.InterOpCollection;
-import com.alibaba.graphscope.gremlin.InterOpCollectionBuidler;
+import com.alibaba.graphscope.gremlin.InterOpCollectionBuilder;
 import com.alibaba.graphscope.gremlin.plugin.script.AntlrToJavaScriptEngineFactory;
 import com.alibaba.pegasus.service.protocol.PegasusClient;
 import com.google.protobuf.InvalidProtocolBufferException;
@@ -149,7 +149,7 @@ public class IrStandardOpProcessor extends StandardOpProcessor {
                 .withResult(o -> {
                     try {
                         if (o != null && o instanceof Traversal) {
-                            InterOpCollection opCollection = (new InterOpCollectionBuidler((Traversal) o)).build();
+                            InterOpCollection opCollection = (new InterOpCollectionBuilder((Traversal) o)).build();
                             IrPlan irPlan = opCollection.buildIrPlan();
 
                             byte[] physicalPlanBytes = irPlan.toPhysicalBytes();
