+++ research/query_service/ir/compiler/src/test/java/com/alibaba/graphscope/gremlin/VertexStepTest.java
@@ -24,7 +24,7 @@ import org.apache.tinkerpop.gremlin.process.traversal.Traversal;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.tinkergraph.structure.TinkerFactory;
-import com.alibaba.graphscope.gremlin.InterOpCollectionBuidler.StepTransformFactory;
+import com.alibaba.graphscope.gremlin.InterOpCollectionBuilder.StepTransformFactory;
 import org.junit.Assert;
 import org.junit.Test;
